Moved sample data folder and split file into data processing and simulation launching
